refactor: Pollering 분리
