+++ payment-service/src/main/java/com/palja/payment_service/application/event/PaymentEventPublisher.java
@@ -1,0 +1,7 @@
+package com.palja.payment_service.application.event;
+
+import com.palja.payment_service.domain.entity.PaymentOutbox;
+
+public interface PaymentEventPublisher {
+    void publish(PaymentOutbox outbox);
+}
